Extract ConfigurationCacheIncludedBuildLogicChangesIntegrationTest
from ConfigurationCacheIncludedBuildChangesIntegrationTest
