+++ subprojects/configuration-cache/src/integTest/groovy/org/gradle/configurationcache/ConfigurationCacheIncludedBuildChangesIntegrationTest.groovy
@@ -26,55 +26,6 @@ import static org.junit.Assume.assumeFalse
 class ConfigurationCacheIncludedBuildChangesIntegrationTest extends AbstractConfigurationCacheIntegrationTest {
 
     @Unroll
-    def "invalidates cache upon change to included #fixtureSpec"() {
-        given:
-        def configurationCache = newConfigurationCacheFixture()
-        def fixture = fixtureSpec.fixtureForProjectDir(file('build-logic'))
-        fixture.setup()
-        settingsFile << """
-            pluginManagement {
-                includeBuild 'build-logic'
-            }
-        """
-        buildFile << """
-            plugins { id('$fixture.pluginId') }
-        """
-
-        when:
-        configurationCacheRunLenient fixture.task
-
-        then:
-        outputContains fixture.expectedOutputBeforeChange
-        configurationCache.assertStateStored()
-
-        when:
-        configurationCacheRunLenient fixture.task
-
-        then:
-        outputContains fixture.expectedOutputBeforeChange
-        configurationCache.assertStateLoaded()
-
-        when:
-        fixture.applyChange()
-        configurationCacheRunLenient fixture.task
-
-        then:
-        outputContains fixture.expectedCacheInvalidationMessage
-        outputContains fixture.expectedOutputAfterChange
-        configurationCache.assertStateStored()
-
-        when:
-        configurationCacheRunLenient fixture.task
-
-        then:
-        outputContains fixture.expectedOutputAfterChange
-        configurationCache.assertStateLoaded()
-
-        where:
-        fixtureSpec << BuildLogicChangeFixture.specs()
-    }
-
-    @Unroll
     def "invalidates cache upon change to #inputName used by included build"() {
 
         assumeFalse(
